Add a check whitespace in job title
Without this change zpr will allow a job to be created that cannot be processed because the title contains whitespace characters. This commit adds a case statement in the zpr::job define type to determine whether the resource title contains whitespace. If it does then the compile will fail with an error reporting that there is whitespace in your resource title, and that this is not supported.