Move tags into an array
Without this change there are unnecessary if statements used to declare
tags to be checked for collection in worker, storage and offsite
classes. Additionally, the ssh key for the worker user is not collected
unless the environment matches $env_tag. This commit moves tags into an array so that it may be
referenced as $tags and allow removal of the if statements. The
ssh_authorized_key resource for the zpr worker user is now collected
regardless of environment when a backup job is declared.