Move ssh forced commands wrapper to python
This commit moves the ssh forced commands wrapper script to python. The
script makes attempts to look for command chaining, and for attempts to
replace the rsync binary. Defaults are set in the zpr user class, but
can be overwritten by declaring a alternate array in zpr params.
Additionally, this commit moves aws.erb out of the resources
subdirectory under templates.