Improve checking of rsync command when collecting files
This commit improves checking of the rsync command when
running a zpr rsync backup job. The original rsync command is sent as an
environment variable via ssh that is then checked against a copy on disk
in the wrapper script executed by ssh forced commands. A number of
changes are made to the module in order to allow for this to
function. Additionally, the zpr_proxy ssh_pub_key fact only attempts to
get the homedir of the zpr_proxy user if the user exists. Without this
change the fact attempts to get the homedir for the zpr_proxy user
whether the user exists or not, and fails.

Exported resource collection is made more specific by requiring the
matching of an arbitrary tag in addition to the variable tag. This
greatly limits the possibility of unintentional resource collection.

The duplicity pkg resource is moved into the zpr module namespace so as
to not have a dependency on code in an implementation repository.

Changes are made to the zpr user resource to ensure consistent home
directory management. Generate ssh key resource is updated to have an
option to create dirs, but not keys, so file resources contained within
the define do not need to be duplicated to ensure consistent creation on
targets.

Lastly, rsync is updated to check if files empty and a valid path. If it
is not then no job is created.