Expose cron schedule for snapshot and duplicity. Add docs
Without this change the schedule for snapshots is hidden from the user
and only run once daily. This is not helpful if you want your backups to
run more often.

Additionally, zpr::job was previously undocumented. This commit adds
documentation for zpr::job. Storage_tag is removed, and server is
replaced by storage.

A readme is updated for the module to provide far more detail.