Replace sshkey type with concat
Without this change the sshkey type can cause unpredictable failures due
to its unreliable nature when dealing with large numbers of keys. Since
there are long standing bugs on this type not being addressed I have
opted to move to concat for management of known hosts instead. The file
resource for known_hosts is removed as a result.

Workers now set source user automatically since a worker without a
source user is not very useful. The parameter is removed from params as
a result.

The server parameter in job is replaced with storage, and storage_tag is
removed.

The rsync_cmd class is refactored to allow restricting a command to a
particular environment if desired and moves resource collection there
instead of the user class.