Support keys other than ecdsa. Fix exclude in rsync
This commit adds support for systems that do not have ecdsa keys. It
will default to a rsa key in that case in order to populate known hosts
on the worker node. Additionally, prior to this commit the rsync
command produced when exclude includes an array does not quote the
values produced. Lastly, support for squeeze is added to the wrapper
script since there is no python2 reference there.