Change share_nfs default and add full share to job
This commit adds the full share parameter to zpr job. This allows a user to get more specific with the zfs share than the zfs share define type currently allows. Additionally, share_nfs is turned on by default for each volume. Without this change a volume is created permitted everyone to access it, and nfs is only enabled if a parent volume has it enabled.