Make ttl for gpg agent configurable and add kill script
This commit adds an option to allow a user to set max ttl and cache ttl
for gpg agent used for zpr offsite backups. Additionally, a job is added
to kill the gpg agent on Sundays at 5pm. This is not currently
configurable, but can be made configurable if needed. This change
ensures that the gpg agent is started each week. Without this change the
agent dies after a few weeks and stops working.