Restructure rsync cron command
This commit updates rsync to use a template for the file resource,
instead of crafting the file from puppet based on variables. This
greatly simplifies the rsync command define type. In order to be
compatible with this method some minor updates to the ssh forced
commands wrapper were necessary and are included here.

Additionally, a check is performed before a backup job is started that
ensures the target volume is mounted over nfs.