Restructure define type data references
This commit restructures the format of define types and their data
references. Prior to this commit there is duplication in the define
types used to build zpr::job. This commit reduces duplication and uses
data from params as defaults. The benefit of this structure is a user is
easily able to set their domain specific defaults and choose what to
expose to users.