Add a timestamp to backup job when queued
This commit adds a timestamp to rsync jobs so they have a timestamp that
can be queried when queued. Without this change there is no way to
determine the time an item enters the queue.

Additionally, this adds a lockfile_progs class and includes it on worker
and offsite roles.