Update forced commands to support python 2.6
This commit updates the zpr forced commands wrapper used for rsync
backups to support python 2.6. Without this change backups fail for
squeeze hosts. Additionally, the checking against a list of static
commands is simplified in order to support the ignore errors flag.
Without this change when ignore errors is set the job fails.

Lastly, the run_backup script used to start a rsync backup job is
updated to remove the lock file on exit. When this is not set and a job
exits due to a failure in the wrapper script, or triggering exit, a
lockfile is left over.