Update usage of timestamp in tsp command
This commit updates the placement of a timestamp in the command passed
to task spooler for duplicity and rsync jobs. Specifically, the command
is moved to the first command executed rather than the last. Without
this change the exit code of the time=$date command printed is returned
by task spooler instead of the exit code of the job.